django-application/requirements.txt
asgiref
certifi
charset-normalizer
Django
django-cors-headers
djangorestframework
idna
requests
sqlparse
tzdata
urllib3
